code/backend/requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
backports.zoneinfo==0.2.1
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.11
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0
drf-yasg==1.21.7
et-xmlfile==1.1.0
geojson==3.1.0
gunicorn==21.2.0
idna==3.6
inflection==0.5.1
kombu==5.3.6
openpyxl==3.1.2
packaging==23.2
pillow==10.3.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing-extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
